refactor(datadog): implement common rate limit middleware
