+++ apps/datadog/src/connectors/datadog/users.test.ts
@@ -1,7 +1,7 @@
 import { http } from 'msw';
 import { describe, expect, test, beforeEach } from 'vitest';
 import { server } from '@elba-security/test-utils';
-import { DatadogError } from '../common/error';
+import { DatadogError } from './common/error';
 import type { DatadogUser } from './users';
 import { getUsers, deleteUser } from './users';
