+++ apps/datadog/src/inngest/middlewares/unauthorized-middleware.test.ts
@@ -3,7 +3,7 @@ import { NonRetriableError } from 'inngest';
 import { db } from '@/database/client';
 import { organisationsTable } from '@/database/schema';
 import { encrypt } from '@/common/crypto';
-import { DatadogError } from '@/connectors/common/error';
+import { DatadogError } from '@/connectors/datadog/common/error';
 import { unauthorizedMiddleware } from './unauthorized-middleware';
 
 const apiKey = 'test-access-token';
